pyproject.toml
[project]
name = "news-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.0.4",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-google-vertexai>=2.0.24",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.7",
    "markdown>=3.8",
    "reportlab>=4.4.1",
    "requests>=2.32.3",
    "streamlit>=1.45.1",
    "tavily-python>=0.7.4",
    "weasyprint>=65.1",
]
